pyproject.toml
[project]
name = "ragas-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "datasets>=4.8.4",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langchain-text-splitters>=1.1.2",
    "langserve>=0.3.3",
    "pandas>=3.0.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "sentence-transformers>=5.4.1",
    "sse-starlette>=3.3.4",
    "uvicorn>=0.44.0",
]
